cuda optimized for sad  parallelization, revisioning project

+++ block_matching/common/include/sad_utils.h
@@ -11,24 +11,22 @@ static inline int computeSAD(const ImageGray& curr, const ImageGray& ref,
     int sad = 0;
     for(int y = 0; y < blockSize; y++) {
         for(int x = 0; x < blockSize; x++) {
-            sad += std::abs(static_cast<int>(curr.at(x1 + x, y1 + y)) - 
-                           static_cast<int>(ref.at(x2 + x, y2 + y))); // cast to int to avoid unsigned char underflow when computing absolute difference
+            sad += std::abs(curr.at(x1 + x, y1 + y)) -(ref.at(x2 + x, y2 + y)); 
         }
     }
     return sad;
 }
 
-// Compute SAD (Sum of Absolute Differences) for RGB color block
-// SAD_RGB = Σ Σ |I_curr[c](x1+i, y1+j) - I_ref[c](x2+i, y2+j)| for all (i,j) in block and all channels c={0,1,2}
+// Compute Sum of Absolute Differences for RGB color blocks
 static inline int computeSADRGB(const ImageColor& curr, const ImageColor& ref,
                                 int x1, int y1, int x2, int y2, int blockSize) {
     int sad = 0;
     for(int y = 0; y < blockSize; y++) {
         for(int x = 0; x < blockSize; x++) {
-            // Sum differences across all 3 color channels (Red, Green, Blue)
+            // Sum differences across all 3 channels (R, G, B)
             for(int c = 0; c < 3; c++) {
                 sad += std::abs(static_cast<int>(curr.at(x1 + x, y1 + y, c)) - 
-                               static_cast<int>(ref.at(x2 + x, y2 + y, c))); // cast to int to avoid unsigned char underflow when computing absolute difference
+                               static_cast<int>(ref.at(x2 + x, y2 + y, c)));
             }
         }
     }
